moved functions to adapter library
